refactored MentionBadge component to ues new approach
