+++ src/components/mention-badge/MentionBadge.constants.ts
@@ -25,10 +25,15 @@ const MENTIONBADGE_PADDING_SIZE_MAP: Record<TMentionBadgeSize, TSpacingToken[]> 
     lg: [75, 50, 25],
 };
 
+const DEFAULT_MENTIONBAGDE_MENTIONS = 1;
+const DEFAULT_MENTIONBAGDE_MENTION_LIMIT = 99;
+
 export {
     MENTIONBADGE_SIZES,
     DEFAULT_MENTIONBADGE_SIZE,
     MENTIONBADGE_HEIGHT_SIZE_MAP,
     MENTIONBADGE_TEXT_SIZE_MAP,
     MENTIONBADGE_PADDING_SIZE_MAP,
+    DEFAULT_MENTIONBAGDE_MENTIONS,
+    DEFAULT_MENTIONBAGDE_MENTION_LIMIT,
 };
