+++ src/components/mention-badge/MentionBadge.stories.mdx
@@ -55,7 +55,7 @@ export const mentionBadgeArgTypes = {
             type: 'color',
         },
     },
-    ...Utils.hideComponentProperties(['width']),
+    ...Utils.hideComponentProperties(['className']),
 };
 
 <Meta component={MentionBadge} title="Components/MentionBadge" parameters={{ status: 'stable' }} />
