+++ src/components/mention-badge/MentionBadge.tsx
@@ -1,81 +1,36 @@
-import styled, { css } from 'styled-components';
-import {
-    FlattenInterpolation,
-    FlattenSimpleInterpolation,
-    ThemedStyledProps,
-} from 'styled-components/ts3.6';
+import React from 'react';
 
-import { Spacing } from '../../foundations/layout';
-import { applyPadding } from '../../foundations/layout/Grid.mixins';
-import { applyShape } from '../../foundations/shape/Shape.mixins';
-import { TTheme } from '../../foundations/theme-provider/themes/theme.types';
-import { applyTextMargin, applyTextStyles } from '../text/Text.mixins';
-
-import MentionBadgeBase from './MentionBadge.base';
 import {
     DEFAULT_MENTIONBADGE_SIZE,
-    MENTIONBADGE_HEIGHT_SIZE_MAP,
-    MENTIONBADGE_PADDING_SIZE_MAP,
-    MENTIONBADGE_TEXT_SIZE_MAP,
+    DEFAULT_MENTIONBAGDE_MENTION_LIMIT,
+    DEFAULT_MENTIONBAGDE_MENTIONS,
 } from './MentionBadge.constants';
 import PMentionBadge from './MentionBadge.props';
+import MentionBadgeRoot from './MentionBadge.root';
 
-const getBadgeColors = ({
-    inverted,
-    background,
-    theme,
-}: ThemedStyledProps<PMentionBadge, TTheme>): FlattenSimpleInterpolation => {
-    if (inverted) {
-        return css`
-            color: ${theme.palette.primary.main};
-            background-color: currentColor;
-            border: 2px solid ${background || theme.background.default};
-        `;
-    }
+const MentionBadge: React.FC<PMentionBadge> = ({
+    mentions = DEFAULT_MENTIONBAGDE_MENTIONS,
+    mentionLimit = DEFAULT_MENTIONBAGDE_MENTION_LIMIT,
+    size = DEFAULT_MENTIONBADGE_SIZE,
+    inverted = false,
+    className,
+    ...rest
+}: PMentionBadge): JSX.Element => {
+    const isUnreadBadge = mentions === 0;
 
-    return css`
-        color: ${theme.text.contrast};
-        background-color: ${theme.background.badge};
-        border: 2px solid ${background || theme.background.default};
-    `;
+    const rootProperties = {
+        className,
+        size,
+        inverted,
+        isUnreadBadge,
+        mentionStringLength: mentions.toString().length,
+    };
+
+    return (
+        <MentionBadgeRoot {...rootProperties} {...rest}>
+            {!isUnreadBadge && (mentions > mentionLimit ? `${mentionLimit}+` : mentions)}
+        </MentionBadgeRoot>
+    );
 };
-
-const MentionBadge = styled(MentionBadgeBase)<PMentionBadge>(
-    ({
-        mentions = 1,
-        size = DEFAULT_MENTIONBADGE_SIZE,
-    }: ThemedStyledProps<PMentionBadge, TTheme>): FlattenInterpolation<
-        ThemedStyledProps<PMentionBadge, TTheme>
-    > => {
-        const isUnreadBadge = mentions === 0;
-        const mentionLength = mentions.toString().length;
-
-        return css`
-            display: flex;
-            align-items: center;
-            justify-content: center;
-
-            ${getBadgeColors};
-
-            ${applyPadding(
-                Spacing.symmetric({
-                    vertical: 0,
-                    horizontal: isUnreadBadge
-                        ? 0
-                        : MENTIONBADGE_PADDING_SIZE_MAP[size][mentionLength - 1],
-                })
-            )};
-
-            ${applyShape({
-                radius: isUnreadBadge ? 'circle' : 'pill',
-                width: isUnreadBadge ? 12 : 'auto',
-                height: MENTIONBADGE_HEIGHT_SIZE_MAP[size],
-            })};
-
-            ${applyTextStyles({ size: MENTIONBADGE_TEXT_SIZE_MAP[size], weight: 'bold' })};
-            ${applyTextMargin({ margin: 'none' })};
-        `;
-    }
-);
 
 export default MentionBadge;
